--- phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/skin/ApplicationSidebarToggleGroupSkin.java
@@ -1,51 +1,0 @@
-package org.phoenicis.javafx.components.skin;
-
-import javafx.scene.control.ToggleButton;
-import org.phoenicis.javafx.components.control.ApplicationSidebarToggleGroup;
-import org.phoenicis.javafx.views.mainwindow.apps.ApplicationsSidebar;
-import org.phoenicis.repository.dto.CategoryDTO;
-
-import java.util.Optional;
-
-import static org.phoenicis.configuration.localisation.Localisation.tr;
-
-/**
- * A {@link SidebarToggleGroupBaseSkin} implementation class used inside the {@link ApplicationsSidebar}
- */
-public class ApplicationSidebarToggleGroupSkin extends
-        SidebarToggleGroupBaseSkin<CategoryDTO, ApplicationSidebarToggleGroup, ApplicationSidebarToggleGroupSkin> {
-    /**
-     * Constructor
-     *
-     * @param control The control belonging to the skin
-     */
-    public ApplicationSidebarToggleGroupSkin(ApplicationSidebarToggleGroup control) {
-        super(control);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected Optional<ToggleButton> createAllButton() {
-        final ToggleButton allCategoryButton = createSidebarToggleButton(tr("All"));
-
-        allCategoryButton.setId("allButton");
-        allCategoryButton.setOnAction(event -> getControl().setNothingSelected());
-
-        return Optional.of(allCategoryButton);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected ToggleButton convertToToggleButton(CategoryDTO category) {
-        final ToggleButton categoryButton = createSidebarToggleButton(category.getName());
-
-        categoryButton.setId(String.format("%sButton", category.getId().toLowerCase()));
-        categoryButton.setOnAction(event -> getControl().setSelectedElement(category));
-
-        return categoryButton;
-    }
-}
