+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/apps/ApplicationsSidebar.java
@@ -4,10 +4,10 @@ import javafx.beans.binding.Bindings;
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.scene.control.CheckBox;
-import org.phoenicis.javafx.components.control.ApplicationSidebarToggleGroup;
-import org.phoenicis.javafx.components.control.ListWidgetSelector;
-import org.phoenicis.javafx.components.control.SearchBox;
-import org.phoenicis.javafx.components.control.SidebarGroup;
+import org.phoenicis.javafx.components.application.control.ApplicationSidebarToggleGroup;
+import org.phoenicis.javafx.components.common.control.ListWidgetSelector;
+import org.phoenicis.javafx.components.common.control.SearchBox;
+import org.phoenicis.javafx.components.common.control.SidebarGroup;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
 import org.phoenicis.javafx.views.common.widgets.lists.CombinedListWidget;
 import org.phoenicis.javafx.views.mainwindow.ui.Sidebar;
