+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/engines/EnginesSidebar.java
@@ -6,10 +6,10 @@ import javafx.collections.transformation.FilteredList;
 import javafx.scene.control.CheckBox;
 import org.phoenicis.engines.dto.EngineCategoryDTO;
 import org.phoenicis.engines.dto.EngineVersionDTO;
-import org.phoenicis.javafx.components.control.EnginesSidebarToggleGroup;
-import org.phoenicis.javafx.components.control.ListWidgetSelector;
-import org.phoenicis.javafx.components.control.SearchBox;
-import org.phoenicis.javafx.components.control.SidebarGroup;
+import org.phoenicis.javafx.components.engine.control.EnginesSidebarToggleGroup;
+import org.phoenicis.javafx.components.common.control.ListWidgetSelector;
+import org.phoenicis.javafx.components.common.control.SearchBox;
+import org.phoenicis.javafx.components.common.control.SidebarGroup;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
 import org.phoenicis.javafx.views.common.widgets.lists.CombinedListWidget;
 import org.phoenicis.javafx.views.mainwindow.ui.Sidebar;
