+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/installations/InstallationsSidebar.java
@@ -3,9 +3,9 @@ package org.phoenicis.javafx.views.mainwindow.installations;
 import javafx.beans.binding.Bindings;
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
-import org.phoenicis.javafx.components.control.InstallationsSidebarToggleGroup;
-import org.phoenicis.javafx.components.control.ListWidgetSelector;
-import org.phoenicis.javafx.components.control.SearchBox;
+import org.phoenicis.javafx.components.installation.control.InstallationsSidebarToggleGroup;
+import org.phoenicis.javafx.components.common.control.ListWidgetSelector;
+import org.phoenicis.javafx.components.common.control.SearchBox;
 import org.phoenicis.javafx.settings.JavaFxSettingsManager;
 import org.phoenicis.javafx.views.common.widgets.lists.CombinedListWidget;
 import org.phoenicis.javafx.views.mainwindow.installations.dto.InstallationCategoryDTO;
